feat(lib): SecretValue
Introduce a struct called `SecretValue` which represents a value within a JSON secret. This allow encapsulating a specific secret value into a strongly-typed value.

Misc: make `resolve()` internal by moving to a non-exported file since jsii does not allow free floating functions.
