+++ packages/cdk8s-plus/src/container.ts
@@ -1,5 +1,5 @@
 import { IConfigMap } from './config-map';
-import { ISecret } from './secret';
+import { ISecret, SecretValue } from './secret';
 import * as k8s from './imports/k8s';
 import { Volume } from './volume';
 
@@ -70,18 +70,27 @@ export class EnvValue {
   }
 
   /**
-   * Create a by reading a specific key inside a secret.
+   * Create an env value by reading a specific key inside a secret.
    *
    * @param secret - The secret.
    * @param key - The key.
    * @param options - Additional options.
    */
   public static fromSecret(secret: ISecret, key: string, options: EnvValueFromSecretOptions = {}): EnvValue {
+    return this.fromSecretValue({ secret, key }, options);
+  }
 
+  /**
+   * Defines an environment value from a secret JSON value.
+   * 
+   * @param secretValue The secret value (secrent + key)
+   * @param options Additional options
+   */
+  public static fromSecretValue(secretValue: SecretValue, options: EnvValueFromSecretOptions = {}): EnvValue {
     const source: k8s.EnvVarSource = {
       secretKeyRef: {
-        name: secret.name,
-        key,
+        name: secretValue.secret.name,
+        key: secretValue.key,
         optional: options.optional,
       },
     };
