+++ packages/cdk8s-plus/src/secret.ts
@@ -18,6 +18,21 @@ export interface ISecret extends IResource {
 }
 
 /**
+ * Represents a specific value in JSON secret.
+ */
+export interface SecretValue {
+  /**
+   * The secret
+   */
+  readonly secret: ISecret;
+
+  /**
+   * The JSON key
+   */
+  readonly key: string;
+}
+
+/**
  * Kubernetes Secrets let you store and manage sensitive information, such as
  * passwords, OAuth tokens, and ssh keys. Storing confidential information in a
  * Secret is safer and more flexible than putting it verbatim in a Pod
